refactor: api-type 위치 수정 및 import 정리
