+++ apps/web/src/app/mypage/page.tsx
@@ -2,10 +2,10 @@
 
 import { Suspense } from 'react';
 
-import useReviewCanvasReady from '@/hooks/use-review-canvas-ready.ts';
-import DisconnectedPage from '@/pages/disconnected-page.tsx';
-import { ReviewServiceProvider } from '@/services/review.tsx';
-import useShop from '@/state/shop.ts';
+import useReviewCanvasReady from '@/hooks/use-review-canvas-ready';
+import DisconnectedPage from '@/pages/disconnected-page';
+import { ReviewServiceProvider } from '@/services/review';
+import useShop from '@/state/shop';
 
 import ConnectedPage from './connected-page';
