+++ apps/web/src/app/replies/[replyID]/edit/page.tsx
@@ -5,11 +5,11 @@ import { useMutation, useQuery } from '@tanstack/react-query';
 import { notFound, useParams } from 'next/navigation';
 
 import CloseButton from '@/components/custom-button';
-import useReviewCanvasReady from '@/hooks/use-review-canvas-ready.ts';
-import type { CreateReplyItemRequest } from '@/services/api-types/review';
-import { useReviewService } from '@/services/review.tsx';
-import useShop from '@/state/shop.ts';
-import { MESSAGE_TYPES, sendMessageToShop } from '@/utils/message.ts';
+import useReviewCanvasReady from '@/hooks/use-review-canvas-ready';
+import type { CreateReplyItemRequest } from '@/models/api-type';
+import { useReviewService } from '@/services/review';
+import useShop from '@/state/shop';
+import { MESSAGE_TYPES, sendMessageToShop } from '@/utils/message';
 
 type PageParams = {
   replyID: string;
