+++ apps/web/src/app/reviews/[reviewID]/delete/page.tsx
@@ -4,11 +4,11 @@ import { useMutation } from '@tanstack/react-query';
 import { useParams } from 'next/navigation';
 
 import DeleteConfirm from '@/components/delete-confrim';
-import useReviewCanvasReady from '@/hooks/use-review-canvas-ready.ts';
-import type { PathInfo } from '@/services/api-types/review';
-import { useReviewService } from '@/services/review.tsx';
-import useShop from '@/state/shop.ts';
-import { MESSAGE_TYPES, sendMessageToShop } from '@/utils/message.ts';
+import useReviewCanvasReady from '@/hooks/use-review-canvas-ready';
+import type { PathInfo } from '@/models/api-type';
+import { useReviewService } from '@/services/review';
+import useShop from '@/state/shop';
+import { MESSAGE_TYPES, sendMessageToShop } from '@/utils/message';
 
 type PageParams = {
   reviewID: string;
