+++ apps/web/src/app/reviews/layout.tsx
@@ -1,6 +1,6 @@
 'use client';
 
-import { ReviewServiceProvider } from '@/services/review.tsx';
+import { ReviewServiceProvider } from '@/services/review';
 
 export default function ReviewLayout({ children }: React.PropsWithChildren) {
   return <ReviewServiceProvider>{children}</ReviewServiceProvider>;
