+++ apps/web/src/components/review/infinite-list.tsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 
 import { useSuspenseInfiniteQuery } from '@tanstack/react-query';
 
-import type { ReviewListFilter, ReviewListSort } from '@/services/api-types/review';
+import type { ReviewListFilter, ReviewListSort } from '@/models/api-type';
 import { useReviewService } from '@/services/review';
 import { useConnectedShop } from '@/state/shop';
