+++ apps/web/src/components/review/my-page/my-paginated-list.tsx
@@ -5,7 +5,7 @@ import { useSuspenseQuery } from '@tanstack/react-query';
 import Pagination from '@/components/pagination';
 import ReviewItem from '@/components/review/item';
 import useMessageToShop from '@/hooks/use-message-to-shop';
-import type { ReviewListFilter, ReviewListSort } from '@/services/api-types/review';
+import type { ReviewListFilter, ReviewListSort } from '@/models/api-type';
 import { useReviewService } from '@/services/review';
 import { useConnectedShop } from '@/state/shop';
 import { MESSAGE_TYPES } from '@/utils/message';
