+++ apps/web/src/components/review/paginated-list.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { useSuspenseQuery } from '@tanstack/react-query';
 
 import useMessageToShop from '@/hooks/use-message-to-shop';
-import type { ReviewListFilter, ReviewListSort } from '@/services/api-types/review';
+import type { ReviewListFilter, ReviewListSort } from '@/models/api-type';
 import { useReviewService } from '@/services/review';
 import { useConnectedShop } from '@/state/shop';
 import { MESSAGE_TYPES } from '@/utils/message';
